+++ application/controllers/Home.php
@@ -17,6 +17,7 @@ class Home extends Application
         $this -> data['airports'] = array (
             "airport" => $this -> flights -> flight_airports ()
 		);
+
         $this->data['no_flights']   = $this -> flights  -> count_flights();
 		$this->data['no_planes']    = $this -> fleets   -> count_planes(); 
 		$this->data['all_flights']  = $this -> flights  -> all();